Wechsle Datentyp von Flugmission.abgesuchteHektar von Integer zu Double.

+++ src/main/java/Flugmissionen/Flugmission.java
@@ -1,21 +1,18 @@
 package Flugmissionen;
 
+import Piloten.Pilot;
+import Reviere.Revier;
 import Wild.Gelege;
-import io.quarkus.hibernate.orm.panache.PanacheEntity;
 import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
 import jakarta.json.bind.annotation.JsonbDateFormat;
-import jakarta.json.bind.annotation.JsonbTransient;
 import jakarta.persistence.*;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
-
-import Piloten.Pilot;
-import Reviere.Revier;
-import jakarta.validation.constraints.NotNull;
 
 @Entity
 public class Flugmission extends PanacheEntityBase
@@ -52,17 +49,19 @@ public class Flugmission extends PanacheEntityBase
 
 	public String kommentar;
 
-	public Integer abgesuchteHektar;
+	public Double abgesuchteHektar;
 
 	@Embedded
 	public Koordinaten koordinaten;
 
 	// Getter und Setter für Hibernate
-	public Revier getRevier() {
+	public Revier getRevier()
+	{
 		return revier;
 	}
 
-	public void setRevier(Revier revier) {
+	public void setRevier(Revier revier)
+	{
 		this.revier = revier;
 	}
 }
